test(snowflake): failing fixtures for PERCENTILE_CONT integer type annotation
PERCENTILE_CONT(fraction) WITHIN GROUP (ORDER BY int/bigint_col) should
return NUMBER(38, 3) in Snowflake, not the sort-column type unchanged.
Empirically verified: NUMBER(p,s) -> NUMBER(min(38,p+3), s+3); INT/BIGINT
are aliases for NUMBER(38,0), so both produce NUMBER(38,3).

+++ tests/fixtures/optimizer/annotate_functions.sql
@@ -5347,7 +5347,7 @@ DOUBLE;
 
 # dialect: snowflake
 PERCENTILE_CONT(0.25) WITHIN GROUP (ORDER BY tbl.int_col);
-INT;
+DECIMAL(38, 3);
 
 # dialect: snowflake
 PERCENTILE_CONT(0.5) WITHIN GROUP (ORDER BY tbl.double_col);
@@ -5355,7 +5355,7 @@ DOUBLE;
 
 # dialect: snowflake
 PERCENTILE_CONT(0.75) WITHIN GROUP (ORDER BY tbl.bigint_col) OVER (PARTITION BY 1);
-BIGINT;
+DECIMAL(38, 3);
 
 # dialect: snowflake
 ARRAY_AGG(tbl.int_col) WITHIN GROUP (ORDER BY tbl.int_col);
